refactor(types): split types into dedicated modules
