+++ src/analyzers/import/resolver.ts
@@ -2,7 +2,8 @@ import { builtinModules } from 'node:module'
 import path from 'node:path'
 import ts from 'typescript'
 
-import type { DependencyEdge, DependencyKind } from '../../types.js'
+import type { DependencyEdge } from '../../types/dependency-edge.js'
+import type { DependencyKind } from '../../types/dependency-kind.js'
 import { isSourceCodeFile } from '../../utils/is-source-code-file.js'
 import { normalizeFilePath } from '../../utils/normalize-file-path.js'
 import type { ModuleReference } from './references.js'
