+++ src/commands/import.ts
@@ -1,7 +1,7 @@
 import { analyzeDependencies } from '../analyzers/import/index.js'
 import { printDependencyTree } from '../output/ascii/import.js'
 import { graphToSerializableTree } from '../output/json/import.js'
-import type { DependencyGraph } from '../types.js'
+import type { DependencyGraph } from '../types/dependency-graph.js'
 import { BaseCommand } from './base.js'
 
 export class ImportCommand extends BaseCommand<DependencyGraph> {
