+++ src/types/color-mode.ts
@@ -1,0 +1,1 @@
+export type ColorMode = boolean | 'auto'
