+++ src/types/dependency-kind.ts
@@ -1,0 +1,1 @@
+export type DependencyKind = 'source' | 'external' | 'builtin' | 'missing'
