+++ src/types/react-symbol-kind.ts
@@ -1,0 +1,1 @@
+export type ReactSymbolKind = 'component' | 'hook'
