+++ src/types/react-usage-edge.ts
@@ -1,0 +1,6 @@
+import type { ReactUsageEdgeKind } from './react-usage-edge-kind.js'
+
+export interface ReactUsageEdge {
+  readonly kind: ReactUsageEdgeKind
+  readonly target: string
+}
