+++ src/types/react-usage-filter.ts
@@ -1,0 +1,3 @@
+import type { ReactSymbolKind } from './react-symbol-kind.js'
+
+export type ReactUsageFilter = 'all' | ReactSymbolKind
