+++ src/types/react-usage-location.ts
@@ -1,0 +1,5 @@
+export interface ReactUsageLocation {
+  readonly filePath: string
+  readonly line: number
+  readonly column: number
+}
